pyproject.toml
[tool.poetry]
name = "fast-api-auth"
version = "0.1.0"
description = "fast api with authentication"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
slowapi = "^0.1.8"


[tool.poetry.group.dev-dependencies.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
respx = "^0.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
